+++ modules/posix/inc/ciaaDevices.h
@@ -59,6 +59,7 @@
 
 /*==================[inclusions]=============================================*/
 #include "ciaaPOSIX_stdint.h"
+#include "ciaaPOSIX_stddef.h"
 
 /*==================[cplusplus]==============================================*/
 #ifdef __cplusplus
@@ -106,10 +107,10 @@ typedef int32_t (*ciaaDevices_close)(ciaaDevices_deviceType const * const device
 typedef int32_t (*ciaaDevices_ioctl)(ciaaDevices_deviceType const * const device, int32_t const request, void * param);
 
 /** \brief read function type */
-typedef ssize_t (*ciaaDevices_read)(ciaaDevices_deviceType const * const device, uint8_t * const buf, uint32_t const nbyte);
+typedef ssize_t (*ciaaDevices_read)(ciaaDevices_deviceType const * const device, uint8_t * const buf, size_t const nbyte);
 
 /** \brief write function type */
-typedef ssize_t (*ciaaDevices_write)(ciaaDevices_deviceType const * const device, uint8_t const * const buf, uint32_t const nbyte);
+typedef ssize_t (*ciaaDevices_write)(ciaaDevices_deviceType const * const device, uint8_t const * const buf, size_t const nbyte);
 
 /** \brief seek function type */
 typedef int32_t (*ciaaDevices_seek)(ciaaDevices_deviceType const * const device, int32_t const offset, uint8_t const whence);
